+++ src/polyester/__init__.py
@@ -3,9 +3,11 @@ __all__ = [
     "RInterpreter",
     "PyInterpreter",
     "RCode",
+    "InterpreterError",
 ]
 
 from polyester.convert_r import RCode
+from polyester.interpreter import InterpreterError
 from polyester.pyinterpreter import PyInterpreter
 from polyester.rinterpreter import RInterpreter
 from polyester.utils import get_interpreter